+++ rocsolver/clients/benchmarks/client.cpp
@@ -31,6 +31,8 @@
 #include "testing_ormxl_unmxl.hpp"
 #include "testing_ormxr_unmxr.hpp"
 #include "testing_potf2_potrf.hpp"
+#include "testing_steqr.hpp"
+#include "testing_sterf.hpp"
 #include "testing_sytxx_hetxx.hpp"
 #include <boost/program_options.hpp>
 
@@ -214,6 +216,10 @@ try
 
         ("rightsv",
          po::value<char>(&argus.right_svect)->default_value('N'),
+         "Only applicable to certain routines")
+
+        ("evect",
+         po::value<char>(&argus.evect)->default_value('N'),
          "Only applicable to certain routines");
     // clang-format on
 
@@ -281,7 +287,11 @@ try
        && argus.right_svect != 'N')
         throw std::invalid_argument("Invalid value for --rightsv");
 
-    // rightsv
+    // evect
+    if(argus.evect != 'V' && argus.evect != 'I' && argus.evect != 'N')
+        throw std::invalid_argument("Invalid value for --evect");
+
+    // workmode
     if(argus.workmode != 'O' && argus.workmode != 'I')
         throw std::invalid_argument("Invalid value for --workmode");
 
@@ -1344,6 +1354,26 @@ try
             testing_sytxx_hetxx<false, true, 1, rocblas_double_complex>(argus);
         else
             throw std::invalid_argument("This function does not support the given --precision");
+    }
+    else if(function == "sterf")
+    {
+        if(precision == 's')
+            testing_sterf<float>(argus);
+        else if(precision == 'd')
+            testing_sterf<double>(argus);
+        else
+            throw std::invalid_argument("This function does not support the given --precision");
+    }
+    else if(function == "steqr")
+    {
+        if(precision == 's')
+            testing_steqr<float, float>(argus);
+        else if(precision == 'd')
+            testing_steqr<double, double>(argus);
+        else if(precision == 'c')
+            testing_steqr<float, rocblas_float_complex>(argus);
+        else if(precision == 'z')
+            testing_steqr<double, rocblas_double_complex>(argus);
     }
     else
         throw std::invalid_argument("Invalid value for --function");